select category bag fixed
