+++ istock/frontend/src/components/product/SelectCategories.js
@@ -1,11 +1,11 @@
 import {useContext, useEffect, useState} from "react";
 import ProductContext from "../../context/ProductContext";
+import ChildListCategory from "./ChildListCategory";
 
 function SelectCategories() {
-    const [products] = useContext(ProductContext);
+    const [products, setProducts] = useContext(ProductContext);
     const [categories, setCategories] = useState([]);
     const [selectedCategory, setSelectedCategory] = useState('');
-    const [categoryResults, setCategoryResults] = useState([]);
 
     async function sendCategoryRequest(selectedCategory) {
         const response = await fetch('http://localhost:8080/product/category', {
@@ -18,9 +18,7 @@ function SelectCategories() {
 
     useEffect(() => {
         if (selectedCategory) {
-            sendCategoryRequest(selectedCategory).then(data => setCategoryResults(data));
-            console.log(selectedCategory);
-            console.log(categoryResults);
+            sendCategoryRequest(selectedCategory).then(data => setProducts(data));
         }
     }, [selectedCategory])
 
@@ -36,32 +34,16 @@ function SelectCategories() {
         sendCategoryTreeRequest().then(data => setCategories(data));
     }, [])
 
-    function ChildList({children, parentName}) {
-        const handleClick = (event, name) => {
-            event.stopPropagation();
-            setSelectedCategory(`${parentName}.${name}`);
-        };
-        if (!children || !children.length) {
-            return null;
-        }
-        return (
-            <ul className={'main-nav'}>
-                {children.map(child => (
-                    <li key={child.name} onClick={(event) => handleClick(event, child.name)}>
-                        {child.name}
-                        <ChildList children={child.children} parentName={`${parentName}.${child.name}`}/>
-                    </li>
-                ))}
-            </ul>
-        );
+    const handleCategoryChange = (currentCategory) => {
+        setSelectedCategory(currentCategory);
     }
 
     return <>
         <div className={'main-nav-block'}>
-            <ChildList children={categories.children} parentName={categories.name} />
+            <ChildListCategory children={categories.children} parentName={categories.name}
+                               handleCategoryChange={handleCategoryChange}/>
         </div>
     </>
-
 }
 
 export default SelectCategories;
